documentsbotorig/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
docxcompose==1.4.0
docxtpl==0.20.1
frozenlist==1.7.0
idna==3.10
Jinja2==3.1.6
lxml==6.0.0
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
setuptools==80.9.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
yarl==1.20.1
